[op][agl][injection] Add application service injection
Introduce an injection web API to obtain the list of applications in
system and launch them, as required by launcher and homescreen.

It is implemented as several parts:
* An injection to expose the APIs in the object navigator.appservice
with methods start(app_id) and getApplications(is_graphical_only).
* A PAL abstraction for obtaining system applications, with an AGL
implementation (that is still empty, but will be populated with
DBUS API).

At this stage, the implementation does not communicate with DBUS yet
and this will be done in next patch.

Bug-AGL: SPEC-4249
Signed-off-by: Jose Dapena Paz <jdapena@igalia.com>